requirements.txt
Flask
mysql-connector-python
redis
bcrypt
pytest
requests
gunicorn